Added radar chart to HomeContent
Added radar chart to HomeContent. This adds getAverageAudioFeatures() which gets the audio feature data for the top 50 songs, finds the averages, and then renders them as a nice radar chart.

For the radar chart I'm using nivo, so note the new dependency.

Also added a new endpoint /getAudioFeatures for the audio feature requests.